fix: make z-score tables float-typed for pandas 3.0 compatibility
compute_zscore() built its result frame with no explicit dtype, which
defaults to object under pandas 3.0; assigning it into the enrichment
table then raised a lossy-set TypeError because pandas 3.0 no longer
silently upcasts object->float64. Declare the z-score frame as float at
construction, and cast the destination table to float before assignment.

Fixes test_cpm_zscore failure surfaced by the newly-revived CI.

+++ phippery/modeling.py
@@ -187,7 +187,10 @@ def zscore(
 
     binning = zscore_pids_binning(beads_ds, data_table, min_Npeptides_per_bin)
 
-    zscore_table = copy.deepcopy(ds[f"{data_table}"].to_pandas())
+    # Cast to float: the source layer may be integer-typed, and pandas 3.0 no
+    # longer silently upcasts on the `.loc` assignment below (it raises on a
+    # lossy set). Z-scores are float-valued, so float is the correct dtype.
+    zscore_table = copy.deepcopy(ds[f"{data_table}"].to_pandas()).astype(float)
     zs_df, mu_df, sigma_df = compute_zscore(
         ds, data_table, binning, lower_quantile_limit, upper_quantile_limit
     )
